Split api tests in separate files
