+++ express/package.json
@@ -9,7 +9,7 @@
     "prod": "node build/index.js",
     "start": "npx tsc && node dist/src/index.js",
     "dev": "npx tsc --watch & nodemon dist/src/index.js",
-    "test": "jest",
+    "test": "jest --runInBand --silent --detectOpenHandles",
     "postman": "ts-node postman/postman.generator.ts",
     "lint": "eslint --ext .ts  .",
     "lint-fix": "eslint --fix . --ext .ts"
